+++ brd-ios/breadwallet/src/Redesign/Scenes/REUSABLE/ItemSelection/ItemSelectionInteractor.swift
@@ -107,62 +107,3 @@ class ItemSelectionInteractor: NSObject, Interactor, ItemSelectionViewActions {
         }
     }
 }
-
-struct FindAddressRequestModel: RequestModelData {
-    let text: String?
-    
-    func getParameters() -> [String: Any] {
-        let params = [
-            "Text": text
-        ]
-        
-        return params.compactMapValues { $0 }
-    }
-}
-
-class FindAddressWorker: BaseApiWorker<FindAddressMapper> {
-    override func getUrl() -> String {
-        guard let urlParams = (requestData as? FindAddressRequestModel),
-              let text = urlParams.text else {
-            return ""
-        }
-                
-        return APIURLHandler.getUrl(LoquateEndpoints.base, parameters: [E.loqateKey, text])
-    }
-    
-    override func getParameters() -> [String: Any] {
-        return requestData?.getParameters() ?? [:]
-    }
-    
-    override func getDecodingStrategy() -> JSONDecoder.KeyDecodingStrategy {
-        return .convertFromPascalCase
-    }
-}
-
-struct FindAddressResponseData: ModelResponse {
-    var items: [FindAddressResponseModel]?
-    
-    struct FindAddressResponseModel: Codable {
-        var text: String?
-    }
-}
-
-struct FindAddress: Model {
-    var text: String?
-}
-
-class FindAddressMapper: ModelMapper<FindAddressResponseData, [FindAddress]> {
-    override func getModel(from response: FindAddressResponseData?) -> [FindAddress]? {
-        return response?.items?.compactMap { .init(text: $0.text)}
-    }
-}
-
-enum LoquateEndpoints: String, URLType {
-    static var baseURL: String = "https://api.addressy.com/Capture/Interactive/Find/v1.1/%@"
-    
-    case base = "json3.ws?Key=%@&Text=%@"
-    
-    var url: String {
-        return String(format: Self.baseURL, rawValue)
-    }
-}
